1. 修正CMakeLists文件中GLM目录-->glm 2. 修改纹理Wrap枚举对应GL枚举，去除gl errors

+++ src/render/BycFBO.h
@@ -24,7 +24,7 @@ public:
 
 private:
     String _name;
-    float _width, _height;
+    int _width, _height;
 
     GLuint _id;
 };
